Optimize memory usage in spectra handling and multiprocessing
- Updated `load_spectra_cache` to read spectra in float32 format, reducing memory overhead.
- Modified `build_line_features_cache` to use 'fork' context for multiprocessing, improving memory efficiency when handling large datasets.
- Adjusted `generate_zone_spectra` to initialize spectra as float32, aligning with memory optimization strategies.
- Enhanced `build_line_tokens.py` to ensure spectra are passed as float32 without unnecessary copying, streamlining data processing.

+++ data/two_zone_pipeline.py
@@ -644,7 +644,9 @@ def generate_zone_spectra(
 ) -> np.ndarray:
     """Synthesise every row of ``sample_table`` (contract-C1 columns)."""
     n_samples = len(sample_table)
-    spectra = np.zeros((n_samples, len(wavelength)))
+    # float32: the spectra are unit-normalised and end up in a float32 tensor,
+    # and float64 doubles both the 8 GB working set and the HDF5 cache.
+    spectra = np.zeros((n_samples, len(wavelength)), dtype=np.float32)
     elements = [c for c in sample_table.columns if c not in _SKIP_COLS]
     conc_mat = sample_table[elements].to_numpy(dtype=np.float64)
     zone_rows = sample_table[list(ZONE_COLUMNS)].to_dict("records")
